requirements.txt
fastapi
uvicorn[standard]
pydantic
pyserial